% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pre.R
\name{data_pre}
\alias{data_pre}
\title{Data preprocessing}
\usage{
data_pre(input.path, type = c("Symbol", "ID"))
}
\arguments{
\item{input.path}{The input file is a tab-delimited text file containing the gene expression profiling data in which each row is a gene and each column is a tumor sample.}

\item{type}{Character string indicating which gene identifier to use when matching.}
}
\value{
A m*n matrix contains the expression values of n matched genes in m tumor samples.
}
\description{
\code{data_pre} This function selects matched genes and scales gene expression values (maximum=1, minimum=0).
}
\details{
The "input.path" refers to the path for input file. The "type" refers to official gene symbol or Entrez gene ID.
}
\examples{
path = system.file("extdata", "example.txt", package = "PreMSI", mustWork = TRUE)
data_pre(path, type = "ID")
}
\author{
Lin Li <cpu_lilin@126.com>, Xiaosheng Wang <xiaosheng.wang@cpu.edu.cn>
}
