% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.R
\docType{data}
\name{feature}
\alias{feature}
\title{Feature}
\format{A data frame with 19 rows and 2 columns. The column names
are as follows:
\describe{
  \item{Symbol}{Official Gene Symbol}
  \item{ID}{Gene Entrez ID} 
}}
\source{
The 19 gene features were selected through pan-cancer 10-fold cross validation model.
}
\usage{
feature
}
\description{
The features are made up of 19 genes.
}
\keyword{datasets}
