% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msi_pre.R
\name{msi_pre}
\alias{msi_pre}
\title{Prediction of microsatellite instability}
\usage{
msi_pre(input_data)
}
\arguments{
\item{input_data}{Tumor samples to be predicted.}
}
\value{
A dataframe with 2 columns:
\item{Sample_ID}{Tumor samples to be predicted.}
\item{MSI_status}{Prediction results ("1" indicates MSI-high, and "0" MSI-low/microsatellite stability).}
}
\description{
\code{msi_pre} Prediction of MSI from gene expression profiling data.
}
\details{
One or multiple tumor samples to be predicted as MSI or non-MSI based on their expression profiles of the gene panel.
}
\examples{
path = system.file("extdata", "example.txt", package = "PreMSIm", mustWork = TRUE)
input_data = data_pre(path, type = "ID")
msi_pre(input_data)
}
\author{
Lin Li <cpu_lilin@126.com>, Xiaosheng Wang <xiaosheng.wang@cpu.edu.cn>
}
