% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreMSI.R
\docType{package}
\name{PreMSIm-package}
\alias{PreMSIm-package}
\alias{PreMSIm}
\title{Prediction of microsatellite instability based on gene expression profiles in cancer}
\description{
Prediction of microsatellite instability by the k-nearest neighbors algorithm (k-NN)
}
